touch up sync_packages.js
- factor up findModuleList from stats.js, to list which
  trace modules are included
- fixup cb when build/* exist
- add DRYRUN env var, to skip `npm publish`
